+++ tests/statebot_enter-then.test.sh
@@ -72,5 +72,10 @@ assert_eq "${PREVIOUS_STATE}" "third" \
 assert_eq "${callback_count}" "${expected_callback_count}" \
   "THEN-callbacks should still really have run ${expected_callback_count} times"
 
+#
+# Cleanup
+#
+statebot_delete "statebot_enter-then"
+
 assert_describe "Can enter directly into states with 'statebot_enter'"
 exit $?